+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 * SIRF/STIR
   - `ScatterEstimation` has extra methods that allow setting masks for the tail-fitting
+  - Error raised in `AcquisitionSensitivityModel.[un]normalise` methods applied to a read-only object.
+
 
 
 ## v3.8.1